pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackebds"
version = "0.4.0"
description = "A security research toolkit for embedded device shellcode generation and vulnerability analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "doudou", email = "your-email@example.com"}
]
keywords = ["security", "embedded", "shellcode", "vulnerability", "pwntools", "IoT"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Operating System :: POSIX :: Linux",
]
requires-python = ">=3.8"
dependencies = [
    "pwntools",
    "colorama",
    "requests",
]

[project.urls]
Homepage = "https://github.com/doudoudedi/"
Repository = "https://github.com/doudoudedi/hackEmbedded"
Issues = "https://github.com/doudoudedi/hackEmbedded/issues"

[project.scripts]
hackebds = "hackebds:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hackebds*"]

[tool.setuptools.package-data]
hackebds = ["data/*.db"]
